+++ packages/angular/src/modules/synergy.module.ts
@@ -6,6 +6,7 @@
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 
 import {
+  SynAlertComponent,
   SynBadgeComponent,
   SynButtonComponent,
   SynButtonGroupComponent,
@@ -25,6 +26,8 @@ import {
   SynOptionComponent,
   SynPopupComponent,
   SynPrioNavComponent,
+  SynProgressBarComponent,
+  SynProgressRingComponent,
   SynRadioButtonComponent,
   SynRadioComponent,
   SynRadioGroupComponent,
@@ -36,6 +39,7 @@ import {
 } from '../components';
 
 const components = [
+  SynAlertComponent,
   SynBadgeComponent,
   SynButtonComponent,
   SynButtonGroupComponent,
@@ -55,6 +59,8 @@ const components = [
   SynOptionComponent,
   SynPopupComponent,
   SynPrioNavComponent,
+  SynProgressBarComponent,
+  SynProgressRingComponent,
   SynRadioComponent,
   SynRadioButtonComponent,
   SynRadioGroupComponent,