+++ packages/components/src/components/progress-ring/progress-ring.custom.styles.ts
@@ -1,0 +1,17 @@
+import { css } from 'lit';
+
+export default css`
+  :host {
+    --size: 120px;
+    --track-width: var(--syn-spacing-x-small);
+  }
+
+  .progress-ring__indicator {
+    stroke-linecap: initial;
+  }
+
+  .progress-ring__label {
+    color: var(--syn-typography-color-text);
+    font: var(--syn-heading-x-large);
+  }
+`;